# contracts/analytics-service.openapi.yaml
openapi: 3.1.0
info:
  title: Smart Campus - Analytics Service API
  description: |
    Hợp đồng REST API cho Analytics Service (Pair 08 & Dashboard).
    Cung cấp API phục vụ Dashboard: lấy báo cáo tổng hợp, thống kê nhiệt độ, lưu lượng và cảnh báo.
    Đồng thời hỗ trợ endpoint nhận event (Queue-Async ingestion) phục vụ kiểm thử container trong Lab 04.
  version: 1.0.0
  contact:
    name: Quỳnh Đinh Trọng (Analytics Team)
    email: dinhtrongquynh240@gmail.com
servers:
  - url: http://localhost:8000
    description: Local or Docker Container Server
  - url: http://localhost:4010
    description: Prism Mock Server
  - url: https://api.smartcampus.dainam.edu.vn
    description: Production Server

tags:
  - name: system
    description: Health and system diagnostics
  - name: reports
    description: Truy vấn báo cáo tổng quan
  - name: metrics
    description: Truy vấn thống kê chuỗi thời gian
  - name: ingestion
    description: Ingestion endpoint cho sự kiện bất đồng bộ

security:
  - bearerAuth: []

paths:
  /health:
    get:
      tags: [system]
      summary: Kiểm tra trạng thái hoạt động của Analytics Service
      description: Endpoint kiểm tra liveness và readiness phục vụ Docker HEALTHCHECK.
      operationId: getHealthStatus
      security: []
      responses:
        "200":
          description: Service hoạt động bình thường
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                healthy:
                  value:
                    status: ok
                    service: analytics-service
                    version: 1.0.0

  /api/v1/analytics/daily-summary:
    get:
      summary: Lấy báo cáo tổng hợp trong ngày
      operationId: getDailySummary
      description: Endpoint cho phép Dashboard lấy báo cáo vận hành theo ngày (định dạng YYYY-MM-DD).
      tags: [reports]
      parameters:
        - name: date
          in: query
          required: true
          description: Ngày cần tra cứu báo cáo (Định dạng YYYY-MM-DD).
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
            example: "2026-08-13"
      responses:
        "200":
          description: Lấy báo cáo thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummaryReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/analytics/metrics/temperature:
    get:
      summary: Lấy dữ liệu nhiệt độ trung bình theo khu vực
      operationId: getTemperatureMetrics
      description: Lấy chuỗi thời gian nhiệt độ để vẽ biểu đồ giám sát IoT.
      tags: [metrics]
      parameters:
        - name: room_id
          in: query
          required: false
          description: Mã phòng/khu vực (Ví dụ ROOM-101).
          schema:
            type: string
            pattern: "^[A-Z0-9-]+$"
            example: "ROOM-101"
      responses:
        "200":
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TemperatureMetric"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/analytics/metrics/access-stats:
    get:
      summary: Lấy thống kê số lượt ra vào theo giờ
      operationId: getAccessStats
      description: Truy vấn mật độ di chuyển qua các cổng để phân tích giờ cao điểm.
      tags: [metrics]
      responses:
        "200":
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessStatMetric"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /api/v1/analytics/metrics/events:
    get:
      summary: Lấy danh sách các sự kiện bất thường và phát hiện chuyển động
      operationId: getAnomalyEvents
      description: Danh sách sự kiện an ninh áp dụng phân trang limit và cursor.
      tags: [metrics]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Thành công
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnomalyEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"

  /api/v1/analytics/events/ingest:
    post:
      summary: Tiếp nhận sự kiện từ các Producer (Queue-Async Simulation)
      operationId: ingestEvent
      description: Nhận event từ IoT Ingestion, Camera Stream, Core Business, hoặc Access Gate theo Event Envelope chuẩn.
      tags: [ingestion]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventEnvelope"
      responses:
        "201":
          description: Event được tiếp nhận và xử lý thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "200":
          description: Event trùng (Idempotent - đã được xử lý trước đó)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Cursor dùng để lấy trang tiếp theo.
      schema:
        type: string
        minLength: 1
        maxLength: 200
        example: "c-20260813-0001"
    Limit:
      name: limit
      in: query
      required: false
      description: Số item tối đa mỗi trang (giới hạn từ 1 đến 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20

  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - service
        - version
      properties:
        status:
          type: string
          example: "ok"
        service:
          type: string
          example: "analytics-service"
        version:
          type: string
          example: "1.0.0"

    DailySummaryReport:
      type: object
      additionalProperties: false
      required:
        - date
        - total_access
        - total_alerts
      properties:
        date:
          type: string
          format: date
          description: Ngày xuất báo cáo
          example: "2026-08-13"
        total_access:
          type: integer
          minimum: 0
          description: Tổng lượt quẹt thẻ ra/vào
          example: 120
        total_alerts:
          type: integer
          minimum: 0
          description: Tổng số cảnh báo an ninh/hệ thống
          example: 5
        avg_temperature:
          type: [number, "null"]
          description: Nhiệt độ trung bình toàn campus.
          example: 30.8

    TemperatureMetric:
      type: object
      additionalProperties: false
      required:
        - room_id
        - recorded_at
      properties:
        room_id:
          type: string
          pattern: "^[A-Z0-9-]+$"
          description: Mã định danh phòng
          example: "ROOM-101"
        avg_temperature:
          type: [number, "null"]
          description: Nhiệt độ trung bình của phòng tại thời điểm đó
          example: 28.5
        sensor_count:
          type: integer
          minimum: 0
          description: Số lượng cảm biến đang hoạt động
          example: 2
        recorded_at:
          type: string
          format: date-time
          description: Mốc thời gian ghi nhận (UTC)
          example: "2026-08-13T08:30:00Z"

    AccessStatMetric:
      type: object
      additionalProperties: false
      required:
        - time_bucket
        - entry_count
        - exit_count
      properties:
        time_bucket:
          type: string
          description: Khoảng thời gian thống kê
          example: "08:00 - 09:00"
        entry_count:
          type: integer
          minimum: 0
          description: Số người vào
          example: 45
        exit_count:
          type: integer
          minimum: 0
          description: Số người ra
          example: 12

    AnomalyEvent:
      type: object
      additionalProperties: false
      required:
        - event_id
        - event_type
        - timestamp
      properties:
        event_id:
          type: string
          description: Mã định danh sự kiện
          example: "a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d"
        event_type:
          type: string
          enum: [MOTION_DETECTED, SECURITY_ALERT, TEMPERATURE_ALERT]
          description: Loại sự kiện bất thường
          example: "SECURITY_ALERT"
        timestamp:
          type: string
          format: date-time
          description: Thời điểm xảy ra
          example: "2026-08-13T09:00:00Z"
        details:
          type: object
          description: Chi tiết sự kiện

    EventEnvelope:
      type: object
      additionalProperties: false
      required:
        - eventId
        - eventType
        - occurredAt
        - correlationId
        - source
        - data
      properties:
        eventId:
          type: string
          description: UUID định danh duy nhất của event
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        eventType:
          type: string
          description: Loại sự kiện nghiệp vụ
          example: "telemetry.ingested"
        occurredAt:
          type: string
          format: date-time
          description: Thời điểm xảy ra sự kiện
          example: "2026-08-13T09:15:00Z"
        correlationId:
          type: string
          description: Correlation ID truy vết luồng
          example: "f1e2d3c4-b5a6-7890-abcd-ef1234567890"
        source:
          type: string
          description: Nguồn phát sinh sự kiện
          example: "iot-ingestion"
        data:
          type: object
          description: Payload chi tiết của event
          example:
            deviceId: "SENSOR-001"
            zoneId: "ZONE-A"
            metric: "temperature"
            value: 38.5
            unit: "celsius"
            sampledAt: "2026-08-13T09:14:58Z"

    IngestResult:
      type: object
      additionalProperties: false
      required:
        - eventId
        - status
        - message
      properties:
        eventId:
          type: string
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        status:
          type: string
          example: "PROCESSED"
        message:
          type: string
          example: "Event processed successfully"

    ProblemDetails:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          description: URI định danh loại lỗi
          example: "https://smart-campus.local/problems/bad-request"
        title:
          type: string
          description: Tiêu đề lỗi
          example: "Bad Request"
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Mô tả chi tiết nguyên nhân lỗi
          example: "Tham số date không đúng định dạng YYYY-MM-DD"
        instance:
          type: string
          description: URI của request gây lỗi
          example: "/api/v1/analytics/daily-summary"

  responses:
    BadRequest:
      description: Dữ liệu gửi lên sai định dạng hoặc thiếu tham số
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: Thiếu hoặc sai Token xác thực Bearer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ValidationError:
      description: Lỗi kiểm tra dữ liệu đầu vào (Validation Error)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    InternalServerError:
      description: Lỗi nội bộ hệ thống Analytics
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"